Divide the map-manager root and the acquisition application module (#711)
The last two production files over the review budget outside the curated
hotspots. Both are a mix of item families, not one indivisible item, so the
earlier reading that they needed a semantic cut was wrong.

| file | before | after | submodules |
| --- | --- | --- | --- |
| map_manager/mod.rs | 2,191 | 255 | grid, terrain, pathfinder, pending_respawn, runtime_state |
| spell_acquisition/application.rs | 1,459 | 147 | validate_plan, validate_post_commit, translate, prepare |

The roots keep `MapManager`, `MapInstance`, the `WorldCreature` impl and the
type alias, and the acquisition request/outcome types. Both ceiling rows are
tightened to the validated figures. Cohesion reviews drop from 65 to 63.

This pass rewrote the divider, because the previous one was wrong in a way
the earlier deliveries never exposed:

- It treated `mod x;` as an item, so a root that declares its submodules
  *before* its imports lost the whole preamble into a group file - the same
  fault #709 hit in a test root. The divider now keeps, by construction,
  every line that is not part of a moved item: docs, attributes, `use`
  statements, `mod` declarations and any unnamed item all stay put.
- It named `const fn power_type_from_u8_like_cpp` as `fn`, because the
  identifier was read after the first keyword rather than after the item
  keyword.
- It wrote a group file without checking whether that path already existed.
  `respawn` and `runtime` are existing modules here, so the first attempt
  overwrote `map_manager/runtime.rs` and produced 376 errors about types
  that had just been deleted. The groups are `pending_respawn` and
  `runtime_state`, and the divider now refuses to overwrite.

Visibility follows the reach each item had, not a rule: an item that was
private becomes `pub(super)` with a private binding in the root, and one that
was `pub(super)` in a crate-root module becomes `pub(crate)` - in
`spell_acquisition/application.rs` it becomes
`pub(in crate::spell_acquisition)`, because that module's parent is not the
crate root. Struct fields a sibling reads were widened from E0616, and the
one binding whose only reader is the test module carries `#[cfg(test)]`.

Item multiset identical per divided file: 104 and 26 items before and after.
wow-world's 3,822 tests pass with no renamed test path, no target gains a
warning, and architecture check, self-test, physical ratchet and hotspot
ratchet pass with every hotspot at or below baseline.

Closes #711
